refactor(cyGraph): separate cyGraph from StyleSheet creation
move StyleSheet creation logic into its own file
