+++ src/components/GraphView.vue
@@ -7,7 +7,8 @@ import { defineComponent } from "vue";
 import cytoscape from "cytoscape";
 import dagre from "cytoscape-dagre";
 import popper from "cytoscape-popper";
-import { makeCyElements, makeCyStylesheets } from "../common/cyGraphFactory";
+import { makeCyElements } from "../common/cyGraphFactory";
+import { makeCyStylesheets } from "../common/cyStyleSheetsFactory";
 import { extendCyPopperElements } from "../common/cyPopperExtender";
 import { Dag } from "../common/dag";
 import { TOP_LEVEL_DAG_ID } from "../common/constants";
